// ts-source/src/enums/roomtype.ts
export enum Roomtype {
    Normal = 1,
    Beachhead = 2,
    Praiseroom = 3,
    Decoration = 4,

    NormalAndNotExpansion = 101,
    NormalWithExpansion = 102,
    NormalWithDismantleTarget = 103,
    NormalWithAcceleratedPraising = 104,
    Expanion = 105,
    My = 106,
    NormalWithPraiseroom = 107,
    BeingAbandoned = 108,
    NormalNotAbandoned = 109,
    NormalUnderSiege = 110
}